chore(hermes): rename core module to kernel so traitreg would not argue
