+++ hermes/bin/src/runtime_extensions/hermes/sqlite/mod.rs
@@ -16,8 +16,8 @@ use crate::{
 };
 
 mod connection;
-mod core;
 mod host;
+mod kernel;
 mod state;
 mod statement;
